fix database and entities

+++ src/main/java/org/acme/syntheticdata/model/OrderLine.java
@@ -3,6 +3,8 @@ package org.acme.syntheticdata.model;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Entity
 @Table(name = "orderline")
@@ -17,7 +19,7 @@ public class OrderLine {
     @ManyToOne
     @JoinColumn(name = "product_id")
     private Product product;
-    private Float amount;
+    private BigDecimal amount;
     private Integer quantity;
     @Column(name = "active", length = 1)
     private Character active = 'Y';
